Little refactorization and fixed free line evaluation
