+++ .gitattributes
@@ -1,22 +1,18 @@
-*.python  text diff=python
+*.python text diff=python
 *.py text diff=python
-
-*.css  	text diff=css
-*.html  text diff=html
-*.tld   text
-*.xml   text
+*.css text diff=css
+*.html text diff=html
+*.tld text
+*.xml text
 *.scss text diff=scss
-
 *.jpg binary
 *.jpeg binary
 *.png binary
-
 *.md text diff=markdown
 *.mdx text diff=markdown
 *.tex text diff=tex
 *.pdf diff=astextplain
 *.PDF diff=astextplain
 *.txt text
-
 .gitignore export-ignore
 .gitattributes export-ignore
