+++ src/main/game_window/engine/Evaluator.py
@@ -2,7 +2,11 @@ from typing import TYPE_CHECKING
 
 from numba import jit
 
-from game_window.engine.StaticEvaluator import StaticEvaluator
+from game_window.engine.static_eval.KingPressure import KingPressure
+from game_window.engine.static_eval.LightPiecesEval import LightPiecesEval
+from game_window.engine.static_eval.PawnEval import PawnEval
+from game_window.engine.static_eval.RookEval import RookEval
+from game_window.engine.static_eval.StaticEvaluator import StaticEvaluator
 
 if TYPE_CHECKING:
     from game_window.Board import Board
@@ -15,11 +19,11 @@ class Evaluator:
 
         material_eval = StaticEvaluator.evaluate_pieces_on_board(board, favor_color)
         center_possession_eval = StaticEvaluator.evaluate_center_possession(board, favor_color)
-        light_dev_eval = StaticEvaluator.evaluate_light_pieces_walked(board, favor_color)
-        king_pressure = StaticEvaluator.evaluate_king_pressure(board, favor_color)
-        bishops = StaticEvaluator.evaluate_bishops(board, favor_color)
-        free_lines = StaticEvaluator.evaluate_free_lines_for_rooks(board, favor_color)
-        chains = StaticEvaluator.evaluate_pawn_chains(board, favor_color)
+        light_dev_eval = LightPiecesEval.evaluate_light_pieces_walked(board, favor_color)
+        king_pressure = KingPressure.evaluate_king_pressure(board, favor_color)
+        bishops = LightPiecesEval.evaluate_bishops(board, favor_color)
+        free_lines = RookEval.evaluate_free_lines_for_rooks(board, favor_color)
+        chains = PawnEval.evaluate_pawn_chains(board, favor_color)
 
         print(f"\tmaterialEval = {material_eval}\n\tcenterPossessionEval = {center_possession_eval}\n\tlightDevEval = {light_dev_eval}\n\tkingPressure = {king_pressure}")
         print(f"\tbishops = {bishops}\n\tfree lines = {free_lines}\n\tpawn chains = {chains}")
